% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_visualization.R
\name{acc_succesion}
\alias{acc_succesion}
\title{Returns a vector with the number of consecutive nodes in each level}
\usage{
acc_succesion(nodes, res = NULL, prev = 0, acc = 0)
}
\arguments{
\item{nodes}{a vetor with the level of each node}

\item{res}{the accumulative results of the sub succesions}

\item{prev}{the level of the previous node processed}

\item{acc}{the accumulator of the index in the current sub succesion}
}
\value{
the vector of sub succesions in each level
}
\description{
This method processes the vector of node levels to get the position of
each node inside the level.
}
\examples{
\dontrun{
levels <- c(1,1,1,2,2,3,4,4,5,5)
levels <- acc_succesion(levels)
levels # equals to c(1,2,3,1,2,1,1,2,1,2)}
}
