% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_visualization.R
\name{acc_successions}
\alias{acc_successions}
\title{Returns a vector with the number of consecutive nodes in each level}
\usage{
acc_successions(nodes, res = NULL, prev = 0, acc = 0)
}
\arguments{
\item{nodes}{a vector with the level of each node}

\item{res}{the accumulative results of the sub successions}

\item{prev}{the level of the previous node processed}

\item{acc}{the accumulator of the index in the current sub successions}
}
\value{
the vector of sub successions in each level
}
\description{
This method processes the vector of node levels to get the position of
each node inside the level. E.g. c(1,1,1,2,2,3,4,4,5,5) turns into 
c(1,2,3,1,2,1,1,2,1,2)
}
\keyword{internal}
