% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{AIC.dbn.fit}
\alias{AIC.dbn.fit}
\title{Calculate the AIC of a dynamic Bayesian network}
\usage{
\method{AIC}{dbn.fit}(object, dt, k = 1)
}
\arguments{
\item{object}{the structure of the network.}

\item{k}{the penalty parameter}

\item{data}{the dataset to calculate the score of the network}

\item{...}{additional parameters for the network scoring}
}
\value{
the prediction results
}
\description{
Generic method for calculating the Akaike information criterion (AIC) of a 
"dbn" S3 object given some data. Calls \code{\link{AIC}} underneath.
}
