% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{AIC.dbn}
\alias{AIC.dbn}
\title{Calculate the AIC of a dynamic Bayesian network}
\usage{
AIC.dbn(object, data, ...)
}
\arguments{
\item{object}{the structure of the network.}

\item{data}{the dataset to calculate the score of the network}

\item{...}{additional parameters for the network scoring}
}
\description{
Generic method for calculating the Akaike information criterion (AIC) of a 
"dbn" S3 object given some data. Calls \code{\link{AIC}} underneath.
}
