% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{all.equal.dbn.fit}
\alias{all.equal.dbn.fit}
\title{Check if two fitted networks are equal to each other}
\usage{
\method{all}{equal.dbn.fit}(target, current, ...)
}
\arguments{
\item{target}{"dbn.fit" object}

\item{current}{the other "dbn.fit" object}

\item{...}{additional parameters}
}
\value{
the prediction results
}
\description{
Generic method for checking the equality of two
"dbn.fit" S3 objects. Calls bnlearn's \code{\link{all.equal}} underneath.
}
