% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{all.equal.dbn}
\alias{all.equal.dbn}
\title{Check if two network structures are equal to each other}
\usage{
\method{all}{equal.dbn}(target, current, ...)
}
\arguments{
\item{target}{"dbn" object}

\item{current}{the other "dbn" object}

\item{...}{additional parameters}
}
\value{
the prediction results
}
\description{
Generic method for checking the equality of two
"dbn" S3 objects. Calls bnlearn's \code{\link{all.equal}} underneath.
}
