% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_inference.R
\name{approximate_inference}
\alias{approximate_inference}
\title{Performs approximate inference forecasting with the GDBN over a data set}
\usage{
approximate_inference(dt, fit, obj_vars, ini, rep, len, num_p)
}
\arguments{
\item{dt}{data.table object with the TS data}

\item{fit}{bn.fit object}

\item{obj_vars}{variables to be predicted}

\item{ini}{starting point in the data set to forecast.}

\item{rep}{number of repetitions to be performed of the approximate inference}

\item{len}{length of the forecast}

\item{num_p}{number of particles to be used by bnlearn}
}
\value{
the results of the forecast
}
\description{
Given a bn.fit object, the size of the net and a data.set,
performs approximate forecasting with bnlearns cpdist function over the 
initial evidence taken from the data set.
}
