% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_inference.R
\name{aprox_prediction_step}
\alias{aprox_prediction_step}
\title{Performs approximate inference in a time slice of the dbn}
\usage{
aprox_prediction_step(fit, variables, particles, n = 50)
}
\arguments{
\item{fit}{bn.fit object}

\item{variables}{variables to be predicted}

\item{particles}{a list with the provided evidence}
}
\value{
the inferred particles
}
\description{
Given a bn.fit object and some variables, performs
particle inference over such variables in the net for a given time slice.
}
