% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_named_vector}
\alias{as_named_vector}
\title{Converts a single row data.table into a named vector}
\usage{
as_named_vector(dt)
}
\arguments{
\item{dt}{a 1 row data.table or a named vector}
}
\value{
a vector with the values and the variable names
}
\description{
Given a single row data.table, convert it into a named vector. Used in the
mvn_inference, to obtain an evidence named vector. For this case, the 
data.table should contain only the evidence that we want to provide. If a
named vector is provided instead of a data.table, nothing will be done and
the named vector will be returned.
}
\keyword{internal}
