% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_inference.R
\name{assign_evidence}
\alias{assign_evidence}
\title{Assign the evidence of the initial time slices in the GDBN}
\usage{
assign_evidence(dt, name, acu, size, slice)
}
\arguments{
\item{dt}{original data.table with "size" rows of initial values}

\item{name}{names of the variables to be evidenced}

\item{acu}{the accumulated evidence to this point in the recursion}

\item{size}{number of time slices}

\item{slice}{current time slice}
}
\value{
the data.table with the evidence and the names
in the first row
}
\description{
Recursive function that returns a data.table
with each variable of the GDBN and its given evidence.
}
