% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{BIC.dbn.fit}
\alias{BIC.dbn.fit}
\title{Calculate the BIC of a dynamic Bayesian network}
\usage{
\method{BIC}{dbn.fit}(object, ...)
}
\arguments{
\item{object}{the fitted network}

\item{...}{additional parameters for the network scoring}
}
\value{
the BIC score of the network
}
\description{
Generic method for calculating the Bayesian information criterion (BIC) of a 
"dbn.fit" S3 object given some data. Calls bnlearn's \code{\link{BIC}} underneath.
}
