% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{BIC.dbn}
\alias{BIC.dbn}
\title{Calculate the BIC of a dynamic Bayesian network}
\usage{
\method{BIC}{dbn}(object, ...)
}
\arguments{
\item{object}{the structure of the network}

\item{...}{additional parameters for the network scoring}
}
\value{
the prediction results
}
\description{
Generic method for calculating the Akaike information criterion (BIC) of a 
"dbn" S3 object given some data. Calls bnlearn's \code{\link{BIC}} underneath.
}
