% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bn_translate_exp}
\alias{bn_translate_exp}
\title{Experimental function that translates a natPosition vector into a DBN network.}
\usage{
bn_translate_exp(ps, ordering_raw, n_arcs, nodes)
}
\arguments{
\item{ps}{a position vector of natural numbers}

\item{ordering_raw}{the ordering of the variables}

\item{n_arcs}{the total number of arcs}

\item{nodes}{the name of all the nodes in the network}
}
\value{
a bn object
}
\description{
Experimental function that translates a natPosition vector into a DBN network.
}
