% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_mu_cpp}
\alias{calc_mu_cpp}
\title{Calculate the mu vector of means of a Gaussian linear network. This is 
the C++ backend of the function.}
\usage{
calc_mu_cpp(fit, order)
}
\arguments{
\item{fit}{a bn.fit object as a Rcpp::List}

\item{order}{a topological ordering of the nodes as a vector of strings}
}
\value{
the map with the nodes and their mu. Returns as a named numeric vector
}
\description{
Calculate the mu vector of means of a Gaussian linear network. This is 
the C++ backend of the function.
}
\keyword{internal}
