% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_sigma_cpp}
\alias{calc_sigma_cpp}
\title{Calculate the sigma covariance matrix of a Gaussian linear network. 
This is the C++ backend of the function.}
\usage{
calc_sigma_cpp(fit, order)
}
\arguments{
\item{fit}{a bn.fit object as a Rcpp::List}

\item{order}{a topological ordering of the nodes as a vector of strings}
}
\value{
the covariance matrix
}
\description{
Calculate the sigma covariance matrix of a Gaussian linear network. 
This is the C++ backend of the function.
}
\keyword{internal}
