% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{$<-.dbn.fit}
\alias{$<-.dbn.fit}
\title{Replacement function for parameters inside DBNs}
\usage{
\method{$}{dbn.fit}(x, name) <- value
}
\arguments{
\item{x}{the fitted network}

\item{name}{name of the node to replace its parameters}

\item{value}{the new parameters}
}
\value{
the modified network
}
\description{
Generic parameter replacement method for "dbn.fit" S3 objects. 
Calls bnlearn underneath.
}
