% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_modification.R
\name{check_time_formatted}
\alias{check_time_formatted}
\title{Checks if the column names of a data.table are time formatted}
\usage{
check_time_formatted(dt)
}
\arguments{
\item{dt}{the data.table to be treated}
}
\value{
TRUE if it is well formatted. FALSE in other case.
}
\description{
This will check if the data.table is properly
time formatted to be folded into more time slices. A data.table is well
formatted when all of its column names end in "_t_0".
}
