% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{coef.dbn.fit}
\alias{coef.dbn.fit}
\title{Extracts the coefficients of a DBN}
\usage{
\method{coef}{dbn.fit}(object, ...)
}
\arguments{
\item{object}{the fitted network}

\item{...}{additional parameters}
}
\value{
the coefficients of the network
}
\description{
Generic method for "dbn.fit" S3 objects. 
Calls bnlearn underneath.
}
