% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_learn.R
\name{create_blacklist}
\alias{create_blacklist}
\title{Creates the blacklist of arcs from a folded data.table}
\usage{
create_blacklist(name, size, acc = NULL, slice = 1)
}
\arguments{
\item{name}{the names of the first time slice, ended in _t_0}

\item{size}{the number of time slices of the net. Markovian 1 would be size 2}

\item{acc}{accumulator of the results in the recursion}

\item{slice}{current time slice that is being processed}
}
\value{
the two column matrix with the blacklisted arcs
}
\description{
This will create the blacklist of arcs that are not
to be learned in the second phase of the dmmhc. This includes arcs
backwards in time or inside time-slices.
}
