% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_causlist_cpp}
\alias{create_causlist_cpp}
\title{Create a causal list from a DBN. This is the C++ backend of the function.}
\usage{
create_causlist_cpp(cl, net, size, ordering)
}
\arguments{
\item{cl}{an initialized causality list}

\item{net}{a dbn object treated as a list of lists}

\item{size}{the size of the DBN}

\item{ordering}{a list with the order of the variables in t_0}
}
\value{
a list with a CharacterVector and a NumericVector
}
\description{
Create a causal list from a DBN. This is the C++ backend of the function.
}
