% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_natcauslist_cpp}
\alias{create_natcauslist_cpp}
\title{Create a natural causal list from a DBN. This is the C++ backend of the function.}
\usage{
create_natcauslist_cpp(cl, net, ordering)
}
\arguments{
\item{cl}{an initialized causality list}

\item{net}{a dbn object treated as a list of lists}

\item{ordering}{a vector with the names of the variables in order}
}
\value{
the natCauslist equivalent to the DBN
}
\description{
Create a natural causal list from a DBN. This is the C++ backend of the function.
}
