% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crop_names_cpp}
\alias{crop_names_cpp}
\title{If the names of the nodes have "_t_0" appended at the end, remove it}
\usage{
crop_names_cpp(names)
}
\arguments{
\item{names}{a vector with the names of the nodes in t_0}
}
\value{
the vector with the names cropped
}
\description{
If the names of the nodes have "_t_0" appended at the end, remove it
}
