% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cte_times_vel_cpp}
\alias{cte_times_vel_cpp}
\title{Multiply a Velocity by a constant real number}
\usage{
cte_times_vel_cpp(k, vl, abs_op, max_op)
}
\arguments{
\item{k}{the constant real number}

\item{vl}{the Velocity's causal list}

\item{abs_op}{the final number of {1,-1} operations}

\item{max_op}{the maximum number of directions in the causal list}
}
\value{
a list with the Velocity's new causal list and number of operations
}
\description{
Multiply a Velocity by a constant real number
}
