% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{degree.bn}
\alias{degree.bn}
\title{Convert a network structure into a model string}
\usage{
\method{degree}{bn}(object, Nodes, ...)
}
\arguments{
\item{object}{a "dbn" object}

\item{Nodes}{which nodes to check}

\item{...}{additional parameters}
}
\value{
the prediction results
}
\description{
Generic method for converting a "dbn" S3 object into a string.
Calls bnlearn's \code{\link{degree}} underneath.
}
