% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{degree.dbn}
\alias{degree.dbn}
\title{Calculates the degree of a list of nodes}
\usage{
\method{degree}{dbn}(object, Nodes, ...)
}
\arguments{
\item{object}{a "dbn" object}

\item{Nodes}{which nodes to check}

\item{...}{additional parameters}
}
\value{
the degree of the nodes
}
\description{
Generic method for calculating the degree of a list of nodes in a
"dbn" S3 object. Calls bnlearn's \code{\link{degree}} underneath.
The degree function is specifically coded to look for either the "bn"
or the "bn.fit" method inside bnlearn, so I have to cast the "dbn" class
into "bn" for it to work.
}
\keyword{internal}
