% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{degree}
\alias{degree}
\title{Convert a network structure into a model string}
\usage{
degree(object, Nodes, ...)
}
\arguments{
\item{object}{a "dbn" object}

\item{Nodes}{which nodes to check}

\item{...}{additional parameters}
}
\value{
the prediction results
}
\description{
Generic method for converting a "dbn" S3 object into a string.
Calls bnlearn's \code{\link{degree}} underneath. I have to redefine the
generic and mask the original for it to work on both bn and dbn objects.
}
