% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_visualization.R
\name{dynamic_ordering}
\alias{dynamic_ordering}
\title{Gets the ordering of a single time slice in a DBN}
\usage{
dynamic_ordering(structure)
}
\arguments{
\item{structure}{the structure of the network.}
}
\value{
the ordered nodes of t_0
}
\description{
This method gets the structure of a DBN, isolates the nodes of a single
time slice and then gives a topological ordering of them.
}
\keyword{internal}
