% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_inference.R
\name{exact_inference_backwards}
\alias{exact_inference_backwards}
\title{Performs exact inference smoothing with the GDBN over a data set}
\usage{
exact_inference_backwards(dt, fit, obj_vars, ini, len, prov_ev)
}
\arguments{
\item{dt}{data.table object with the TS data}

\item{fit}{bn.fit object}

\item{obj_vars}{variables to be predicted. Should be in the oldest time step}

\item{ini}{starting point in the data set to smooth}

\item{len}{length of the smoothing}

\item{prov_ev}{variables to be provided as evidence in each forecasting step. Should be in the oldest time step}
}
\value{
the results of the smoothing
}
\description{
Given a bn.fit object, the size of the net and a data set,
performs exact smoothing over the initial evidence taken from the data set.
Take notice that the smoothing is done backwards in time, as opposed to
forecasting.
}
\keyword{internal}
