% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_inference.R
\name{exact_prediction_step}
\alias{exact_prediction_step}
\title{Performs exact inference in a time slice of the dbn}
\usage{
exact_prediction_step(fit, variables, evidence)
}
\arguments{
\item{fit}{list with the mu and sigma of the MVN model}

\item{variables}{variables to be predicted}

\item{evidence}{a list with the provided evidence}
}
\value{
a list with the predicted mu and sigma
}
\description{
Given a bn.fit object and some variables, performs
exact MVN inference over such variables in the net for a given time slice.
}
\keyword{internal}
