% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_visualization.R
\name{expand_time_nodes}
\alias{expand_time_nodes}
\title{Extends the names of the nodes in t_0 to t_(max-1)}
\usage{
expand_time_nodes(name, acc, max, i)
}
\arguments{
\item{name}{the names of the nodes in the t_0 slice}

\item{acc}{accumulator of the resulting names in the recursion}

\item{max}{number of time slices in the net}

\item{i}{current slice being processed}
}
\value{
the extended names
}
\description{
This method extends the names of the nodes to the given maximum and
mantains the order of the nodes in each slice, so as to plotting
the nodes in all slices relative to their homonyms in the first slice.
}
