% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_learn.R
\name{fit_dbn_params}
\alias{fit_dbn_params}
\title{Fits a markovian n DBN model}
\usage{
fit_dbn_params(net, f_dt, ...)
}
\arguments{
\item{net}{the structure of the DBN}

\item{f_dt}{a folded data.table}

\item{...}{aditional parameters for bnlearn's bn.fit function}
}
\value{
the fitted net
}
\description{
Fits the parameters of the DBN via MLE or BGE. The "mu" vector of means 
and the "sigma" covariance matrix are set as attributes of the dbn.fit 
object for future exact inference.
}
