% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{fitted.dbn.fit}
\alias{fitted.dbn.fit}
\title{Extracts the fitted values of a DBN}
\usage{
\method{fitted}{dbn.fit}(object, ...)
}
\arguments{
\item{object}{the fitted network}

\item{...}{additional parameters}
}
\value{
the fitted values of the network
}
\description{
Generic method for "dbn.fit" S3 objects. 
Calls bnlearn underneath.
}
