% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_modification.R
\name{fold_dt}
\alias{fold_dt}
\title{Widens the dataset to take into account the t previous time slices}
\usage{
fold_dt(dt, size)
}
\arguments{
\item{dt}{the data.table to be treated}

\item{size}{number of time slices to unroll. Markovian 1 would be size 2}
}
\value{
the extended data.table
}
\description{
This will widen the dataset to put the t previous time slices
in each row, so that it can be used to learn temporal arcs in the second
phase of the dmmhc.
}
\examples{
data(motor)
size <-  3
dt <- fold_dt(motor, size)
}
