% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_modification.R
\name{fold_dt_rec}
\alias{fold_dt_rec}
\title{Widens the dataset to take into account the t previous time slices}
\usage{
fold_dt_rec(dt, n_prev, size, slice = 1)
}
\arguments{
\item{dt}{the data.table to be treated}

\item{n_prev}{names of the previous time slice}

\item{size}{number of time slices to unroll. Markovian 1 would be size 2}

\item{slice}{the current time slice being treated. Should not be modified
when first calling.}
}
\value{
the extended data.table
}
\description{
This will widen the dataset to put the t previous time slices
in each row, so that it can be used to learn temporal arcs in the second
phase of the dmmhc. Recursive version not exported, the user calls from the
handler 'fold_dt'
}
