% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init_cl_cpp}
\alias{init_cl_cpp}
\title{Initialize the nodes vector}
\usage{
init_cl_cpp(n_nodes)
}
\arguments{
\item{n_nodes}{number of receiving nodes}
}
\value{
a list with the randomly initialized particles
}
\description{
Initialize the vector in C++
}
