% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init_list_cpp}
\alias{init_list_cpp}
\title{Initialize the particles}
\usage{
init_list_cpp(nodes, size, n_inds)
}
\arguments{
\item{nodes}{the names of the nodes}

\item{size}{the size of the DBN}

\item{n_inds}{the number of particles}
}
\value{
a list with the randomly initialized particles
}
\description{
Initialize the particles
}
