% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initialize_cl_cpp}
\alias{initialize_cl_cpp}
\title{Create a causality list and initialize it}
\usage{
initialize_cl_cpp(ordering, size)
}
\arguments{
\item{ordering}{a list with the order of the variables in t_0}

\item{size}{the size of the DBN}
}
\value{
a causality list
}
\description{
Create a causality list and initialize it
}
