% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{mean.dbn.fit}
\alias{mean.dbn.fit}
\title{Average the parameters of multiple dbn.fit objects with identical structures}
\usage{
\method{mean}{dbn.fit}(object, ..., k)
}
\arguments{
\item{object}{the fitted network}

\item{...}{additional parameters}

\item{weights}{optional weights}
}
\value{
the averaged parameters
}
\description{
Generic method for "dbn.fit" S3 objects. 
Calls bnlearn underneath.
}
