% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_learn.R
\name{merge_nets}
\alias{merge_nets}
\title{Merges and replicates the arcs in the static BN into all the time-slices
in the DBN}
\usage{
merge_nets(net0, netCP1, size, acc = NULL, slice = 1)
}
\arguments{
\item{net0}{the structure of the static net}

\item{netCP1}{the state transition net}

\item{size}{the number of time slices of the net. Markovian 1 would be size 2}

\item{acc}{accumulator of the results in the recursion}

\item{slice}{current time slice that is being processed}
}
\value{
the merged nets
}
\description{
This will join the static net and the state transition
net by replicating the arcs in the static net in all the time slices.
}
