% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_exact_inference.R
\name{mvn_inference}
\alias{mvn_inference}
\title{Performs inference over a multivariate normal distribution}
\usage{
mvn_inference(sigma, mu, evidence)
}
\arguments{
\item{sigma}{the covariance matrix}

\item{mu}{the mean vector}

\item{evidence}{values and names of the variables given as evidence}
}
\value{
the inferred value of the rest of the variables in the net
}
\description{
Performs inference over a multivariate normal distribution given some 
evidence. After converting a Gaussian linear network to its MVN form, this
kind of inference can be performed.
}
