% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_inference.R
\name{name_variables}
\alias{name_variables}
\title{Gets a vector with the names of the variables in the wanted slices.}
\usage{
name_variables(variables, acu, size, slice)
}
\arguments{
\item{variables}{variables in the time slice t0}

\item{acu}{accumulated vector of names}

\item{size}{number of time slices}

\item{slice}{current time slice}
}
\value{
the vector of variable names
}
\description{
Recursive function that returns the names of the
variables in all the slices between "slice" and "size".
}
