% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nat_cte_times_vel_cpp}
\alias{nat_cte_times_vel_cpp}
\title{Multiply a Velocity by a constant real number}
\usage{
nat_cte_times_vel_cpp(k, vl, vl_neg, abs_op, max_size)
}
\arguments{
\item{k}{the constant real number}

\item{vl}{the Velocity's positive causal list}

\item{vl_neg}{the Velocity's negative causal list}

\item{abs_op}{the final number of {1,-1} operations}

\item{max_size}{the maximum size of the network}
}
\value{
the new total number of operations
}
\description{
Multiply a Velocity by a constant real number
}
