% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nat_pos_minus_pos_cpp}
\alias{nat_pos_minus_pos_cpp}
\title{Subtracts two natPositions to obtain the natVelocity that transforms ps1 into ps2}
\usage{
nat_pos_minus_pos_cpp(ps1, ps2, vl, vl_neg)
}
\arguments{
\item{ps1}{the first position's causal list}

\item{ps2}{the second position's causal list}

\item{vl}{the natVelocity's positive causal list}

\item{vl_neg}{the natVelocity's negative causal list}
}
\value{
the velocity's causal lists by reference and the number of operations by return
}
\description{
Subtracts two natPositions to obtain the natVelocity that transforms ps1 into ps2
}
