% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nat_pos_plus_vel_cpp}
\alias{nat_pos_plus_vel_cpp}
\title{Add a velocity to a position}
\usage{
nat_pos_plus_vel_cpp(cl, vl, vl_neg, n_arcs)
}
\arguments{
\item{cl}{the position's causal list}

\item{vl}{the velocity's positive causal list}

\item{vl_neg}{velocity's negative causal list}

\item{n_arcs}{number of arcs present in the position. Remainder: can't return integers by reference, they get casted to 1 sized vectors}
}
\value{
the new position by reference and the new number of arcs by return
}
\description{
Add a velocity to a position
}
