% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{natcl_to_arc_matrix_cpp}
\alias{natcl_to_arc_matrix_cpp}
\title{Create a matrix with the arcs defined in a causlist object}
\usage{
natcl_to_arc_matrix_cpp(cl, ordering, rows)
}
\arguments{
\item{cl}{a causal list}

\item{ordering}{a list with the order of the variables in t_0}

\item{rows}{number of arcs in the network}
}
\value{
a StringMatrix with the parent nodes and the children nodes
}
\description{
Create a matrix with the arcs defined in a causlist object
}
