% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{nodes<-.bn.fit}
\alias{nodes<-.bn.fit}
\title{Relabel the names of the nodes of a BN or a DBN}
\usage{
\method{nodes}{bn.fit}(object) <- value
}
\arguments{
\item{object}{a "bn.fit" object}

\item{value}{a list with the new names}
}
\value{
the modified object
}
\description{
Generic method for renaming the nodes in a "bn.fit" S3 object.
Calls bnlearn's \code{\link{nodes<-}} underneath.
}
\keyword{internal}
