% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{nodes.dbn}
\alias{nodes.dbn}
\title{Returns a list with the names of the nodes of a BN or a DBN}
\usage{
\method{nodes}{dbn}(object, ...)
}
\arguments{
\item{object}{a "dbn" object}

\item{...}{additional parameters}
}
\value{
the names of the nodes
}
\description{
Generic method for obtaining the names of the nodes in "dbn" S3 object.
Calls bnlearn's \code{\link{nodes}} underneath.
The nodes function is specifically coded to look for either the "bn"
or the "bn.fit" method inside bnlearn, so I have to cast the "dbn" class
into "bn" for it to work.
}
