% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{nodes}
\alias{nodes}
\title{Returns a list with the names of the nodes of a BN or a DBN}
\usage{
nodes(object, ...)
}
\arguments{
\item{object}{a "bn", "dbn", "bn.fit" or "dbn.fit" object}

\item{...}{additional parameters}
}
\value{
the names of the nodes
}
\description{
Generic method for obtaining the names of the nodes in a BN or a DBN.
Calls bnlearn's \code{\link{nodes}} underneath. 
I have to redefine the generic and mask the original for it to work on both 
bn and dbn objects without the user having to import bnlearn.
}
