% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nodes_gen_exp}
\alias{nodes_gen_exp}
\title{Generates the names of the nodes in t_0 and in all the network}
\usage{
nodes_gen_exp(ordering, size)
}
\arguments{
\item{ordering}{the names of the variables}

\item{size}{the desired size of the dbn}
}
\value{
a dictionary with the variable names in t_0 and in all other time slices
}
\description{
Given the names of the desired variables, this function generates the names
of the variables in a DBN without needing a previous dataset. It's just a 
wrapper around the 'fold_dt' function.
}
