% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ordering_gen_exp}
\alias{ordering_gen_exp}
\title{Generates the names of n variables.}
\usage{
ordering_gen_exp(n)
}
\arguments{
\item{n}{the total number of variables desired}
}
\value{
a character vector with the variable names
}
\description{
Given the total number of variables, this function generates a character
vector with variables named "Xi", where i is a number in the interval [0,n-1]
}
