% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_visualization.R
\name{plot.dbn}
\alias{plot.dbn}
\title{Plots a dynamic Bayesian network}
\usage{
\method{plot}{dbn}(x, ...)
}
\arguments{
\item{x}{the structure of the network.}

\item{...}{additional parameters for the visualization of a DBN}
}
\description{
Generic method for plotting the "dbn" S3 objects. Calls 
\code{\link{plot_dynamic_network}} underneath.
}
