% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_visualization.R
\name{plot_network}
\alias{plot_network}
\title{Plots a Bayesian networks or a dynamic Bayesian network}
\usage{
plot_network(structure, ...)
}
\arguments{
\item{structure}{the structure or fit of the network.}

\item{...}{additional parameters for the visualization of a DBN}
}
\description{
Selects the appropriate function depending on the class of the object
}
