% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pos_plus_vel_cpp}
\alias{pos_plus_vel_cpp}
\title{Add a velocity to a position}
\usage{
pos_plus_vel_cpp(cl, vl, n_arcs)
}
\arguments{
\item{cl}{the position's causal list}

\item{vl}{the velocity's causal list}

\item{n_arcs}{number of arcs present in the position}
}
\value{
a list with the modified position and the new number of arcs
}
\description{
Add a velocity to a position
}
