% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_inference.R
\name{predict.dbn.fit}
\alias{predict.dbn.fit}
\title{Predicts the values in t_0 of every row in a dataset with a dynamic Bayesian network}
\usage{
\method{predict}{dbn.fit}(object, dt, obj_nodes, verbose = T, look_ahead = F)
}
\arguments{
\item{object}{the fitted bn}

\item{dt}{the test data set}

\item{obj_nodes}{the nodes that are going to be predicted. They are all predicted at the same time}

\item{verbose}{if TRUE, displays the metrics and plots the real values against the predictions}

\item{look_ahead}{boolean that defines whether or not the values of the variables in t_0 should be used when predicting, even if they are not present in obj_nodes. This decides if look-ahead bias is introduced or not.}
}
\value{
the prediction results
}
\description{
Generic method for predicting a dataset with a "dbn.fit" S3 objects. Calls 
\code{\link{predict_dt}} underneath.
}
