% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_inference.R
\name{predict_bn}
\alias{predict_bn}
\title{Performs inference over a fitted GBN}
\usage{
predict_bn(fit, evidence)
}
\arguments{
\item{fit}{the fitted bn}

\item{evidence}{values of the variables used as evidence for the net}
}
\value{
the mean of the particles for each row
}
\description{
Performs inference over a Gaussian BN. It's thought to be used in a map for
a data.table, to use as evidence each separate row.
}
\examples{
res <- dt_test[, predict_bn(fit, .SD), by = 1:nrow(dt_test)]

}
