% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{print.dbn.fit}
\alias{print.dbn.fit}
\title{Print method for "dbn.fit" objects}
\usage{
\method{print}{dbn.fit}(x, ...)
}
\arguments{
\item{x}{the "dbn.fit" object}

\item{...}{additional parameters}
}
\description{
Generic print method for "dbn.fit" S3 objects. Calls bnlearn's print underneath
}
