% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{print.dbn}
\alias{print.dbn}
\title{Print method for "dbn" objects}
\usage{
\method{print}{dbn}(x, ...)
}
\arguments{
\item{x}{the "dbn" object}

\item{...}{additional parameters}
}
\description{
Generic print method for "dbn" S3 objects. Calls bnlearn's print underneath
}
