% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{rbn.dbn.fit}
\alias{rbn.dbn.fit}
\title{Simulates random samples from a fitted DBN}
\usage{
\method{rbn}{dbn.fit}(object, ..., weights)
}
\arguments{
\item{object}{the fitted network}

\item{...}{additional parameters}

\item{weights}{optional weights}
}
\value{
the averaged parameters
}
\description{
Generic method for "dbn.fit" S3 objects. 
Calls bnlearn's \code{\link{rbn}} underneath.
}
