% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recount_arcs_exp}
\alias{recount_arcs_exp}
\title{Experimental function that recounts the number of arcs in the position}
\usage{
recount_arcs_exp(ps)
}
\arguments{
\item{ps}{a position vector of natural numbers}
}
\value{
the number of arcs
}
\description{
Experimental function that recounts the number of arcs in the position
}
