% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{residuals.dbn.fit}
\alias{residuals.dbn.fit}
\title{Returns the residuals from fitting a DBN}
\usage{
\method{residuals}{dbn.fit}(object, ...)
}
\arguments{
\item{object}{the fitted network}

\item{...}{additional parameters}
}
\value{
the residuals of fitting the network
}
\description{
Generic method for "dbn.fit" S3 objects. 
Calls bnlearn underneath.
}
