% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{score.bn}
\alias{score.bn}
\title{Computes the score of a BN or a DBN}
\usage{
\method{score}{bn}(object, ...)
}
\arguments{
\item{object}{a "bn" object}

\item{...}{additional parameters}
}
\value{
the score of the network
}
\description{
Generic method for computing the score of a "bn" S3 object.
Calls bnlearn's \code{\link{score}} underneath.
}
\keyword{internal}
