% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{score.dbn}
\alias{score.dbn}
\title{Computes the score of a BN or a DBN}
\usage{
\method{score}{dbn}(object, ...)
}
\arguments{
\item{object}{a "dbn" object}

\item{...}{additional parameters}
}
\value{
the score of the network
}
\description{
Generic method for computing the score of a "dbn" S3 object.
Calls bnlearn's \code{\link{score}} underneath.
The nodes function is specifically coded to look for either the "bn"
or the "bn.fit" method inside bnlearn, so I have to cast the "dbn" class
into "bn" for it to work.
}
\keyword{internal}
