% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{score}
\alias{score}
\title{Computes the score of a BN or a DBN}
\usage{
score(object, ...)
}
\arguments{
\item{object}{a "bn" or "dbn" object}

\item{...}{additional parameters}
}
\value{
the score of the network
}
\description{
Generic method for computing the score of a BN or a DBN.
Calls bnlearn's \code{\link{nodes}} underneath. 
I have to redefine the generic and mask the original for it to work on both 
bn and dbn objects without the user having to import bnlearn.
}
