% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{sigma.dbn.fit}
\alias{sigma.dbn.fit}
\title{Returns the standard deviation of the residuals from fitting a DBN}
\usage{
\method{sigma}{dbn.fit}(object, ...)
}
\arguments{
\item{object}{the fitted network}

\item{...}{additional parameters}
}
\value{
the standard deviation residuals of fitting the network
}
\description{
Generic method for "dbn.fit" S3 objects. 
Calls bnlearn underneath.
}
