% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_modification.R
\name{time_rename}
\alias{time_rename}
\title{Renames the columns in a data.table so that they end in '_t_0'}
\usage{
time_rename(dt)
}
\arguments{
\item{dt}{the data.table to be treated}
}
\value{
the renamed data.table
}
\description{
This will rename the columns in a data.table so that
they end in '_t_0', which will be needed when folding the data.table. If
any of the columns already ends in '_t_0', a warning will be issued and
no further operation will be done.
}
\examples{
data("motor")
dt <- time_rename(motor)
}
