% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_transform.R
\name{transform_bnfit}
\alias{transform_bnfit}
\title{Calculate the mu mean vector and the sigma covariance matrix of a Gaussian 
linear network and return a list with the original fit, mu and sigma.}
\usage{
transform_bnfit(fit)
}
\arguments{
\item{fit}{a bn.fit object}
}
\value{
a list with the bnfit, mu and sigma
}
\description{
Calculate the mu mean vector and the sigma covariance matrix of a Gaussian 
linear network and return a list with the original fit, mu and sigma.
}
