% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trunc_geom}
\alias{trunc_geom}
\title{Geometric distribution sampler truncated to a maximum}
\usage{
trunc_geom(p, max)
}
\arguments{
\item{p}{the parameter of the geometric distribution}

\item{max}{the maximum value allowed to be sampled}
}
\value{
the sampled value
}
\description{
A geometric distribution sampler with probability 'p' restricted to values
inside [1, max]. Because of this restriction, very low values of 'p' 
coupled with low 'max' return increasingly uniform populations in 
the interval [1, max].
}
