% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vel_plus_vel_cpp}
\alias{vel_plus_vel_cpp}
\title{Add two Velocities}
\usage{
vel_plus_vel_cpp(vl1, vl2, abs_op)
}
\arguments{
\item{vl1}{the first Velocity's causal list}

\item{vl2}{the second Velocity's causal list}

\item{abs_op}{the final number of {1,-1} operations}
}
\value{
a list with the Velocity's causal list and the number of operations
}
\description{
Add two Velocities
}
