% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_learn.R
\name{add_attr_to_fit}
\alias{add_attr_to_fit}
\title{Adds the mu vector and sigma matrix as attributes to the bn.fit or dbn.fit object}
\usage{
add_attr_to_fit(fit)
}
\arguments{
\item{fit}{a fitted bn or dbn}
}
\value{
the fitted net with attributes
}
\description{
Adds the mu vector and sigma matrix as attributes to the bn.fit or dbn.fit 
object to allow performing exact MVN inference on both cases.
}
