% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_transform.R
\name{calc_mu}
\alias{calc_mu}
\title{Calculate the mu vector of means of a Gaussian linear network. Front end
of a C++ function.}
\usage{
calc_mu(fit)
}
\arguments{
\item{fit}{a bn.fit object}
}
\value{
a named numeric vector of the means of each variable
}
\description{
Calculate the mu vector of means of a Gaussian linear network. Front end
of a C++ function.
}
