% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_transform.R
\name{calc_sigma}
\alias{calc_sigma}
\title{Calculate the sigma covariance matrix of a Gaussian linear network. 
Front end of a C++ function.}
\usage{
calc_sigma(fit)
}
\arguments{
\item{fit}{a bn.fit object}
}
\value{
a numeric covariance matrix of the nodes
}
\description{
Calculate the sigma covariance matrix of a Gaussian linear network. 
Front end of a C++ function.
}
