% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_learning_psoho.R
\name{Causlist}
\alias{Causlist}
\title{This file contains all the classes needed for the PSOHO structure learning 
algorithm. It was implemented as an independent package in
https://github.com/dkesada/PSOHO and then merged into dbnR. All the original
source files are merged into one to avoid bloating the R/ folder of the
package.}
\arguments{
\item{ordering}{a vector with the names of the nodes in t_0}

\item{size}{number of timeslices of the DBN}
}
\value{
A new 'causlist' object
}
\description{
Constructor of the 'Causlist' class
}
\details{
The classes are now not exported because the whole algorithm is 
encapsulated inside the package and only the resulting dbn structure is
wanted. As a result, many security checks have been omitted.
R6 class that defines causal lists in the PSO

The causal lists will be the base of the positions and the velocities
in the pso part of the algorithm.
}
\section{Fields}{

\describe{
\item{\code{cl}}{List of causal units}

\item{\code{size}}{Size of the DBN}

\item{\code{ordering}}{String vector defining the order of the nodes in a timeslice}
}}

