% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{degree.bn}
\alias{degree.bn}
\title{Calculates the degree of a list of nodes}
\usage{
\method{degree}{bn}(object, Nodes, ...)
}
\arguments{
\item{object}{a "dbn" object}

\item{Nodes}{which nodes to check}

\item{...}{additional parameters}
}
\value{
the degree of the nodes
}
\description{
Generic method for calculating the degree of a list of nodes in a
"bn" S3 object. Calls bnlearn's \code{\link{degree}} underneath.
}
