% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_compat.R
\name{degree}
\alias{degree}
\title{Calculates the degree of a list of nodes}
\usage{
degree(object, Nodes, ...)
}
\arguments{
\item{object}{a "bn", "dbn", "bn.fit" or "dbn.fit" object}

\item{Nodes}{which nodes to check}

\item{...}{additional parameters}
}
\value{
the degree of the nodes
}
\description{
#' Generic method for calculating the degree of a list of nodes in a
BN or a DBN. Calls bnlearn's \code{\link{degree}} underneath. 
I have to redefine the generic and mask the original for it to work on both 
bn and dbn objects without the user having to import bnlearn.
}
