% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_same_cycle}
\alias{filter_same_cycle}
\title{Filter the instances in a data.table that have values of different ids in each row}
\usage{
filter_same_cycle(f_dt, size, id_var)
}
\arguments{
\item{f_dt}{folded data.table}

\item{size}{the size of the data.table}

\item{id_var}{the variable that labels each individual instance of the time series}
}
\value{
the filtered dataset
}
\description{
Given an id variable that labels the different instances of a time series
inside a dataset, discard the rows that have values from more than 1 id.
}
