% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_learn.R
\name{fit_dbn_params}
\alias{fit_dbn_params}
\title{Fits a markovian n DBN model}
\usage{
fit_dbn_params(f_dt, net)
}
\arguments{
\item{f_dt}{a folded data.table}

\item{net}{the structure of the DBN}
}
\value{
the fitted net
}
\description{
Fits the parameters of the DBN via MLE or BGE.
}
