% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_inference.R
\name{forecast_ts}
\alias{forecast_ts}
\title{Performs forecasting with the GDBN over a data set}
\usage{
forecast_ts(dt, fit, size, obj_vars, ini = 1, len = dim(dt)[1] - 1,
  rep = 1, num_p = 50, print_res = TRUE, plot_res = TRUE)
}
\arguments{
\item{dt}{data.table object with the TS data}

\item{fit}{bn.fit object}

\item{size}{number of time slices of the net}

\item{ini}{starting point in the data set to forecast.}

\item{len}{number of points of the TS to forecast}

\item{rep}{number of times to repeat the forecasting}
}
\value{
the results of the forecast
}
\description{
Given a bn.fit object, the size of the net and a data.set,
performs a forecast over the initial evidence taken from the data set.
}
