% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_visualization.R
\name{plot_network}
\alias{plot_network}
\title{Plots a Bayesian networks in a hierarchical way}
\usage{
plot_network(structure)
}
\arguments{
\item{structure}{the structure of the network.}
}
\description{
Calculates the levels of each node and then plots them in a hierarchical
layout in visNetwork.
}
