% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_visualization.R
\name{plot_network}
\alias{plot_network}
\title{Plots a Bayesian networks in a hierarchical way}
\usage{
plot_network(structure)
}
\arguments{
\item{net}{the structure of the network.}

\item{size}{the number of time slices of the net. Markovian 1 would be size 2}

\item{acc}{accumulator of the results in the recursion}

\item{slice}{current time slice that is being processed}
}
\value{
the visualzation of the net
}
\description{
Calculates the levels of each node and then plots them in a hierarchical
layout in visNetwork.
}
