% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pos_minus_pos_cpp}
\alias{pos_minus_pos_cpp}
\title{Substracts two Positions to obtain the Velocity that transforms one into the other}
\usage{
pos_minus_pos_cpp(cl, ps, vl)
}
\arguments{
\item{cl}{the first position's causal list}

\item{ps}{the second position's causal list}

\item{vl}{the Velocity's causal list}
}
\value{
a list with the Velocity's causal list and the number of operations
}
\description{
Substracts two Positions to obtain the Velocity that transforms one into the other
}
