% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_inference.R
\name{predict.dbn.fit}
\alias{predict.dbn.fit}
\title{Predicts the values in t_0 of every row in a data set with a dynamic Bayesian network}
\usage{
\method{predict}{dbn.fit}(object, ...)
}
\arguments{
\item{object}{a "dbn.fit" object}

\item{...}{additional parameters for the inference process}
}
\value{
the prediction results
}
\description{
Generic method for predicting a data set with a "dbn.fit" S3 objects. Calls 
\code{\link{predict_dt}} underneath.
}
