% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_inference.R
\name{predict.dbn.fit}
\alias{predict.dbn.fit}
\title{Performs inference in every row of a dataset with a DBN}
\usage{
\method{predict}{dbn.fit}(object, ...)
}
\arguments{
\item{object}{a "dbn.fit" object}

\item{...}{additional parameters for the inference process}
}
\value{
a data.table with the prediction results
}
\description{
Generic method for predicting a dataset with a "dbn.fit" S3 objects. Calls 
\code{\link{predict_dt}} underneath.
}
