% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_inference.R
\name{predict_dt}
\alias{predict_dt}
\title{Performs inference over a test data set with a GBN}
\usage{
predict_dt(fit, dt, obj_nodes, verbose = T)
}
\arguments{
\item{fit}{the fitted bn}

\item{verbose}{if TRUE, displays the metrics and plots the real values against the predictions}

\item{dt_test}{the test data set}
}
\value{
the prediction results
}
\description{
Performs inference over a test data set, plots the results
and gives metrics of the accuracy of the results.
}
