% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shift_values}
\alias{shift_values}
\title{Function that moves the window of values backwards in a folded dataset row}
\usage{
shift_values(f_dt, row)
}
\arguments{
\item{f_dt}{a folded dataset}

\item{row}{the row that is going to be processed}
}
\value{
the shifted row
}
\description{
Move the values in t_0, t_1, ..., t_n-1 in a folded dataset row to
t_1, t_2, ..., t_n. This is useful to predict the values in the last row
of a folded dataset
}
